+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - (**Breaking Change**) Error handling has been reworked to be clearer and more consistent. This includes renaming the
   fields `code` and `number` to `statusCode` and `errorCode`. Error handling has been updated so that exception data is
   more consistent across API resources (see [issue 506](https://github.com/freshbooks/freshbooks-nodejs-sdk/issues/506)).
+- Support for `/accounting` error responses for newer API versions.
 - Drop support for node 12.x
 - Upgrade axios